Handle reexports
Determine all transitively (potential) reexported modules, tag their ABI hash files and pass these on
(using `--abi`) to the ghc wrapper. The wrapper will mark modules as used in the dep-file (non-home
modules are always marked as used).

Direct dependencies of a module are passed as hidden inputs and will always cause a rebuild
of the current module when the ABI hash file changes.

Change-Id: I5fb2db299611c8c8487c3304ffae224197c7ff42